requirements.txt
torch
torchvision
tensorboard
scipy
numpy<2.0.0
matplotlib>=3.7.0
opencv_python
openai
rouge
openpyxl
seaborn
rouge_score
brokenaxes
python-dotenv==1.0.1
pyyaml
bitsandbytes>=0.39.0
transformers==4.41.2
datasets>=2.16.0,<=2.21.0
accelerate>=0.30.1,<=0.33.0
peft>=0.11.1,<=0.12.0
trl>=0.8.6,<=0.9.6
gradio>=4.0.0
pandas>=2.0.0
einops
sentencepiece
tiktoken
protobuf
uvicorn
pydantic
fastapi
sse-starlette
fire
packaging
zhipuai
argparse